djangoAPI/requirements.txt
django==4.2.4
requests==2.31.0
djangorestframework==3.12.0
drf-yasg ==1.20.0